Update to v1.0.16. Fix disposal of migration sessions. (#530)
Problem
When using KEYS option for slot migration, the migration session removal isn't disposing the session object. This leads to an out of memory exception when migrating a large number of keys.

Fix

Set slots correctly for the migrate operation. This helps to ensure the slots are correctly reset on removal of migration session.
Also, removal of multiple outstanding session objects when node is removed for the FORGET command, will dispose each of those session objects.